poprawki

+++ examples/conway/main.c
@@ -20,30 +20,30 @@ int main()
   const uint8_t addr = 0x71;
   matrix8x8_init(addr);
   
-  int sum[10][10];
-  int cell[8][8];
-  int newsum[10][10];
+  uint8_t sum[10][10];
+  uint8_t cell[8][8];
+  uint8_t newsum[10][10];
 
   for(;;)
     {
-      for(int i=0; i<10; i++)
-	for(int j=0; j<10; j++)
+      for(uint8_t i=0; i<10; i++)
+	for(uint8_t j=0; j<10; j++)
 	  sum[i][j]=0;
-      for(int i=0; i<8; i++)
-	for(int j=0; j<8; j++)
+      for(uint8_t i=0; i<8; i++)
+	for(uint8_t j=0; j<8; j++)
 	  cell[i][j]=0;
       
       matrix8x8_clear();
       matrix8x8_writedisplay(addr);
 
-      for(int x=0; x<8; x++)
-	for(int y=0; y<8; y++)
+      for(uint8_t x=0; x<8; x++)
+	for(uint8_t y=0; y<8; y++)
 	  if((rand()&1) + (rand()&1) == 0)
 	    {
 	      cell[x][y]=1;
 	      matrix8x8_drawpixel(x, y, 1);
-	      for(int i=0; i<3; i++)
-		for(int j=0; j<3; j++)
+	      for(uint8_t i=0; i<3; i++)
+		for(uint8_t j=0; j<3; j++)
 		  if(i != 1 || j != 1)
 		    sum[x+i][y+j]++;
 	    }
@@ -52,14 +52,14 @@ int main()
       for(;;)
 	{
 	  
-	  for(int i=0; i<10; i++)
-	    for(int j=0; j<10; j++)
+	  for(uint8_t i=0; i<10; i++)
+	    for(uint8_t j=0; j<10; j++)
 	      newsum[i][j]=sum[i][j];
 	  
 	  int fixpoint = 1;
 	  
-	  for(int x=0; x<8; x++)
-	    for(int y=0; y<8; y++)
+	  for(uint8_t x=0; x<8; x++)
+	    for(uint8_t y=0; y<8; y++)
 	      if(cell[x][y])
 		{
 		  // zywa
@@ -69,8 +69,8 @@ int main()
 		      cell[x][y]=0;
 		      fixpoint=0;
 		      matrix8x8_drawpixel(x, y, 0);
-		      for(int i=0; i<3; i++)
-			for(int j=0; j<3; j++)
+		      for(uint8_t i=0; i<3; i++)
+			for(uint8_t j=0; j<3; j++)
 			  if(i != 1 || j != 1)
 			    newsum[x+i][y+j]--;
 		    }
@@ -84,15 +84,15 @@ int main()
 		      cell[x][y]=1;
 		      fixpoint=0;
 		      matrix8x8_drawpixel(x, y, 1);
-		      for(int i=0; i<3; i++)
-			for(int j=0; j<3; j++)
+		      for(uint8_t i=0; i<3; i++)
+			for(uint8_t j=0; j<3; j++)
 			  if(i != 1 || j != 1)
 			    newsum[x+i][y+j]++;
 		    }
 		}
 	  
-	  for(int i=0; i<10; i++)
-	    for(int j=0; j<10; j++)
+	  for(uint8_t i=0; i<10; i++)
+	    for(uint8_t j=0; j<10; j++)
 	      sum[i][j]=newsum[i][j];
 	  
 	  matrix8x8_writedisplay(addr);
